refactor: split W&B report vocabulary
